[trust-dns-proto] Refactor tests
Abstract test functions and make them pulic, which allows the tests to be run
on different platforms.
