+++ crates/proto/src/tests/mod.rs
@@ -1,0 +1,10 @@
+//! Tests for TCP and DUP stream and client
+
+mod tcp;
+mod udp;
+
+pub use self::tcp::tcp_client_stream_test;
+pub use self::tcp::tcp_stream_test;
+pub use self::udp::next_random_socket_test;
+pub use self::udp::udp_client_stream_test;
+pub use self::udp::udp_stream_test;
