+++ crates/proto/src/udp/udp_client_stream.rs
@@ -345,147 +345,22 @@ impl SingleUseUdpSocket {
 mod tests {
     #![allow(clippy::dbg_macro, clippy::print_stdout)]
 
-    use futures::future;
     #[cfg(not(target_os = "linux"))]
     use std::net::Ipv6Addr;
     use std::net::{IpAddr, Ipv4Addr};
-    use tokio;
-
-    use super::*;
-    use crate::op::Message;
+    use tokio::{net::UdpSocket as tokioUdpSocket, runtime::Runtime};
+    use crate::tests::udp_client_stream_test;
 
     #[test]
     fn test_udp_client_stream_ipv4() {
-        udp_client_stream_test(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)))
+        let io_loop = Runtime::new().expect("failed to create tokio runtime");
+        udp_client_stream_test::<tokioUdpSocket, Runtime>(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), io_loop)
     }
 
     #[test]
     #[cfg(not(target_os = "linux"))] // ignored until Travis-CI fixes IPv6
     fn test_udp_client_stream_ipv6() {
-        udp_client_stream_test(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)))
-    }
-
-    fn udp_client_stream_test(server_addr: IpAddr) {
-        use crate::op::Query;
-        use crate::rr::rdata::NULL;
-        use crate::rr::{Name, RData, Record, RecordType};
-        use std::str::FromStr;
-        use tokio::runtime;
-
-        // use env_logger;
-        // env_logger::try_init().ok();
-
-        let succeeded = std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false));
-        let succeeded_clone = succeeded.clone();
-        std::thread::Builder::new()
-            .name("thread_killer".to_string())
-            .spawn(move || {
-                let succeeded = succeeded_clone;
-                for _ in 0..15 {
-                    std::thread::sleep(std::time::Duration::from_secs(1));
-                    if succeeded.load(std::sync::atomic::Ordering::Relaxed) {
-                        return;
-                    }
-                }
-
-                panic!("timeout");
-            })
-            .unwrap();
-
-        let server = std::net::UdpSocket::bind(SocketAddr::new(server_addr, 0)).unwrap();
-        server
-            .set_read_timeout(Some(std::time::Duration::from_secs(5)))
-            .unwrap(); // should receive something within 5 seconds...
-        server
-            .set_write_timeout(Some(std::time::Duration::from_secs(5)))
-            .unwrap(); // should receive something within 5 seconds...
-        let server_addr = server.local_addr().unwrap();
-
-        let mut query = Message::new();
-        let test_name = Name::from_str("dead.beef").unwrap();
-        query.add_query(Query::query(test_name.clone(), RecordType::NULL));
-        let test_bytes: &'static [u8; 8] = b"DEADBEEF";
-        let send_recv_times = 4;
-
-        let test_name_server = test_name;
-        // an in and out server
-        let server_handle = std::thread::Builder::new()
-            .name("test_udp_client_stream_ipv4:server".to_string())
-            .spawn(move || {
-                let mut buffer = [0_u8; 512];
-
-                for i in 0..send_recv_times {
-                    // wait for some bytes...
-                    debug!("server receiving request {}", i);
-                    let (len, addr) = server.recv_from(&mut buffer).expect("receive failed");
-                    debug!("server received request {} from: {}", i, addr);
-
-                    let request =
-                        Message::from_vec(&buffer[0..len]).expect("failed parse of request");
-                    assert_eq!(*request.queries()[0].name(), test_name_server.clone());
-                    assert_eq!(request.queries()[0].query_type(), RecordType::NULL);
-
-                    let mut message = Message::new();
-                    message.set_id(request.id());
-                    message.add_queries(request.queries().to_vec());
-                    message.add_answer(Record::from_rdata(
-                        test_name_server.clone(),
-                        0,
-                        RData::NULL(NULL::with(test_bytes.to_vec())),
-                    ));
-
-                    // bounce them right back...
-                    let bytes = message.to_vec().unwrap();
-                    debug!("server sending response {} to: {}", i, addr);
-                    assert_eq!(
-                        server.send_to(&bytes, addr).expect("send failed"),
-                        bytes.len()
-                    );
-                    debug!("server sent response {}", i);
-                    std::thread::yield_now();
-                }
-            })
-            .unwrap();
-
-        // setup the client, which is going to run on the testing thread...
-        let mut io_loop = runtime::Runtime::new().unwrap();
-
-        // the tests should run within 5 seconds... right?
-        // TODO: add timeout here, so that test never hangs...
-        // let timeout = Timeout::new(Duration::from_secs(5));
-        let stream = UdpClientStream::with_timeout(server_addr, Duration::from_millis(500));
-        let mut stream: UdpClientStream<tokio::net::UdpSocket> =
-            io_loop.block_on(stream).ok().unwrap();
-        let mut worked_once = false;
-
-        for i in 0..send_recv_times {
-            // test once
-            let response_future = io_loop.block_on(future::lazy(|cx| {
-                stream.send_message(DnsRequest::new(query.clone(), Default::default()), cx)
-            }));
-            println!("client sending request {}", i);
-            let response = match io_loop.block_on(response_future) {
-                Ok(response) => response,
-                Err(err) => {
-                    println!("failed to get message: {}", err);
-                    continue;
-                }
-            };
-            println!("client got response {}", i);
-
-            let response = Message::from(response);
-            if let RData::NULL(null) = response.answers()[0].rdata() {
-                assert_eq!(null.anything().expect("no bytes in NULL"), test_bytes);
-            } else {
-                panic!("not a NULL response");
-            }
-
-            worked_once = true;
-        }
-
-        succeeded.store(true, std::sync::atomic::Ordering::Relaxed);
-        server_handle.join().expect("server thread failed");
-
-        assert!(worked_once);
+        let io_loop = Runtime::new().expect("failed to create tokio runtime");
+        udp_client_stream_test::<tokioUdpSocket, Runtime>(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)), io_loop)
     }
 }
